+++ lib/ui/screens/overview_screen.dart
@@ -66,6 +66,7 @@ class OverviewScreen extends StatelessWidget {
     return DonatedFoodList(
       title: context.l10n!.lastDonated,
       itemsLimit: 5,
+      alwaysShowTitle: false,
     );
   }
 }